Add dummy return dict when niter=0 in tclean
Prior to this commit, tclean would only create a return dict when
niter > 0, since the return dict was generated by the deconvolution
module.

This commit introduces a function in `task_tclean.py` that generates the
return dict prior to initializing the minor cycle, which saves on
compute time and will eventually allow for the CASA GUI interactive
clean to generate sensible convergence plots.

At the moment the return dict is a placeholder, which will be populated
in subsequent commits.