Add summaryminor to tclean return dict for niter=0
In the case of `niter=0`, tclean now returns a full dictionary, with the
`summaryminor` keyword being populated using values from the residual
image on disk.

This functionality is duplicating what exists within the `deconvolver`
module, and is implemented for the specific case of `niter=0`. This has
been done within Python, inside of `task_tclean.py` utilizing the
`casatools.image` and `casatools.coordsys` modules.