Add keys for fullsummary support in tclean
The return dict in `task_tclean.py` now returns the full set of keys as
required by tclean when fullsummary=True. The function also respects the
fullsummary=True/False choice.

This has been tested to work with multiple channels, but it will not do
the right thing for multiple fields.