Add multifield support for tclean return dict
The return dict from tclean as defined in `task_tclean.py` now also
supports multiple fields. This has been tested to work, along with the
multi-channel use case.

Previously the code was interrogating the `bparm` dictionary to figure
out the image name and other parameters. The `bparm` dict doesn't keep
track of the extra fields specified in `outlierfile`. So the code now
uses the `paramList()` object to query for all the parameters, and the
bookkeeping for the `outlierfile` comes through that class.

This commit also removes the unused parameter `field_id` from the
`fillSummaryMinor` function call.