Add tests for single-field return dictionary
Added tests to check the structure of the return dictionary for the
single-field case, both cube and MFS imaging. MFS imaging testing
also probes hogbom and mt-mfs use cases.

At present, the values within the return dictionary are not checked for
correctness, just that the dictionary itself has the right structure.