CAS-13960 : Fix case of existing model image
In the event of an `niter=0` run with an existing model image on disk,
`tclean` would still report `modelFlux=0` from the return dict. This
commit implements the change to read the model flux from disk (if it
exists) and return the correct model flux, and returns zero if the image
does not exist.