CAS-13401: Fix for beam header in full-stokes MTMFS
This fixes the issue of full Stokes MT-MFS images not containing the
restoring beam information.

The bug was because internally MT-MFS reads in a subimage of the
restored image (per channel/polarization plane, per term) in order to
perform the residual normalization math and add the model. The beam
header was being set correctly within this sub-image per plane, but not
in the final image written to disk.

The fix takes the beam header from the final subimage, and applies that
to all planes of the final image written to disk. This is not a problem
at the moment, because mt-mfs has only one frequency channel, and the
restoring beam for all Stokes should be identical.

The fix will need to be revisited in the event of using MT-MFS to image
cubes (with very coarse channelization for example).