CAS-13401 : Generalized fix from previous commit
The previous commit assumed that the restoring beam is identical for
every channel and polarization, and applied the last restoring beam to
all the planes.

The new fix correctly stores the restoring beam per channel/polarization
and applies it to the final restored image. This allows for the use case
of multi-frequency mt-mfs, where the beam can reasonably change across
the different channels.