CAS-14184 : Add initial ImagingDict class
The ImagingDict class is meant to be an API that can accumulate the
return dictionaries from multiple tclean + deconvolve calls, and keep a
track of the total number of iterations, the peak residual per loop.

At the moment, every time a new dict is passed in, it
appends/accumulates a set of pre-defined keys, and only retains the
final value of the remaining keys.