CAS-13901 : Fix asp behavior in deconvolve
This change forces asp to always generate per-plane scales, regardless
of the input specmode. When running asp on a cube from within
deconvolve, specmode defaults to "mfs" which would cause asp to generate
scales for the first plane, and re-use it for all subsequent planes of
the cube.

This is not right, and also diverges from the behavior in tclean. This
change fixes that.

There is also a minor fix in an output log message of AspMatrixCleaner.