CAS-14184: Use deconvolve for minorcycle in gclean
Split the tclean call into using tclean for the major cycles and
deconvolve for the minor cycles. This also makes use of the
`ImagingDict` class in `imagerhelpers.imager_return_dict` that allows
muiltiple return dictionaries to be appended.

The `ReturnDict` and `ImagingDict` class within `imager_return_dict`
have been merged, and the changes have been propogated through
`task_tclean` and `task_deconvolve`. There was a lot of duplicate
functionality between the two classes which has now been consolidated.

The details of the masking behaviour still need to be more thoroughly
tested - namely the interaction between a user-specified input mask and
what is drawn within interactive clean.