CAS-14201 : Modify tclean(niter=0) to respect mask
In the niter=0 case, task_tclean.py has been modified to now respect the
input mask if present. If there is no input mask present, a default mask
is generated (similar to the niter > 0 case) -- this is a
change from previous behaviour where tclean did not create a mask for
niter=0. This change now brings it in line with the behaviour of
task_deconvolve, that similarly creates a mask for niter=0

Currently the behaviour for a newly generated mask with niter=0 follows
that of niter>0 - i.e., if there is no mask present, and the "default"
is generated, for the purposes of computing statistics it is assumed
that the mask is filled with ones rather than zeros.