CAS-14201 : Roll back mask creation
This rolls back the mask creation performed in the previous commit. As
of this commit - `tclean(niter=0)` will not create a mask, i.e., it
maintains existing behavior.

The statistics calculated for niter=0 will still respect an input mask
if provided - i.e., the statistics reported in the return dictionary
will respect the input mask. If no mask is provided, the stats are
calculated over the entire image.