CAS-14184 : Refactor gclean
This commit is a fairly substantial refactor of _gclean.py to correctly
use tclean + deconvolve for the major and minor cycles respectively.
This refactor comes after discussions with Urvashi, and most of the
iteration control is based off the return dictionaries.

The `ImagingDict` class has a few additional functions for iteration
control, such as determining the cyclethreshold per minor cycle.

There are still several small bugs/issues to be resolved, but as of this
commit the basic imaging seems to work.