CAS-14201 : Add check for mask validity
If a mask is provided to `tclean(niter=0)` it now checks for the
existence _and_ validity of a mask image on disk. This fixes a bug where
a string/textfile passed into the `mask` parameter of tclean would
trigger it to try and open the mask with `ia.open`.

Now if the mask image is not present on disk, it gracefully fails.

Note that at this point it is not desirable to change the behaviour of
tclean(niter=0) to create a mask if none is present on the disk.