CAS-14184 : Add parameter checking to gclean
The iteration control parameters (nmajor, niter etc.) that get passed in
via the GUI are currently being checked in casagui.interactiveclean.
This check should happen inside gclean.

The check makes sure that the iteration control params are of the right
type (int, float etc.) and have values that are in an allowable range
(i.e., niter = -100 is not allowed)

This check happens both during initialization of gclean (inside
__init__) as well as in the update() function when new parameters come
in via the GUI.