CAS-14184 : More fixes to parameter checking
Parameter checking is no longer in a function, but rather just inside
gclean.update. There are two reasons for this :

(a) The parameters that come into __init__ already get verified in
interactiveclean, so we don't need to do an additional check at the end.
That leaves the only place to check parameters in gclean.update

(b) In order for the function to work, the internal variables (like
self._niter, self._nmajor etc.) will take on the incorrect value and
then only be checked. Which is not ideal, since we want to prevent these
variables from ever getting to an incorrect state.

The current code does error checking in-situ, so the variables never
enter an incorrect state, but will rather throw and error and go back to
the user to fix things.