CAS-14184 : More fixes for parameter checking
There was a bug checking the threshold value, where the incorrect value
is actually set internally prior to being checked. While this does not
affect interactive functionality, non-interactive tests were failing.

cyclefactor was being tested as an integer, but really it can be a
floating point value. This has been fixed as well.