CAS-13901 : Fix asp tests in deconvolve tests
Edited the passing thresholds for the asp tests in test_task_deconvolve.
Used a smalled gain value (gain=0.2) similar to test_task_tclean, and
the passing criterion are based on the values for a single major cycle
in test_task_tclean.

Also enabled asp testing on Darwin - with a lower gain it is expected to
be more stable across different systems.