CAS-14184 : Bugfix for niter=0 peakres calculation
If no mask was passed in for the niter=0 case, the `ImagingDict`
class set the `mask` parameter to -1 (instead of the pixel mask).
peakRes was always calcualted as `np.amax(image_pixels * mask_pixels)`
so when mask = -1 it picked the highest negative as the peakRes.

This commit fixes that bug, and has been confirmed to pass the
previously failing task_tclean tests.

Additionally, removed a check in test_task_tclean that asserted
len(summaryminor) == 0 for niter = 0.