CAS-14184 : Add check for user updated mask
The user can update/change the mask in between major cycles. Previously
gclean was invoking `deconvolve` to do this, but now there is an
additional function that uses `imstat` to check if the peakresidual has
changed because of a mask update in between major cycles.

This has implications on whether to continue deconvolution or not, and
therefore needs to be done before deconvolve is called.