CAS-14184 : Fixes to gclean tests + masking bugfix
If a static user mask was provided in the input (e.g.,
circle[[50px,50px],10px]) the final mask-update deconvolve step was
breaking. It did not break for auto-masking or any other type of input
mask. This has been fixed by setting mask = '' in the last deconvolve
call, which still does the right thing for auto-masking etc.

Further, the default masking behaviour changed as of a few commits ago.
This was not reflected in the test cases, so they were all failing. The
tests have been updated to reflect this new behaviour as well.