CAS-14184 : Fix test to account for previous change
The "test_mask_preserve_input_zero_mask" test in test_task_deconvolve
was failing because of the wrong stopcode. As of the previous commit the
stopcode is now over-written by the niter=0 dict, rather than what is
written by the C++ code. This has no operation impact apart from the
specific niter=0 case, and the stopcode for this test had to be changed
to reflect that.