CAS-13901 : Remove asp niter check in test_deconvolve
The niter passing criterion has been removed from the asp test within
test_task_deconvolve. This brings it in line with the same test within
test_task_tclean as well.

The issue is that the LBFGS minimizer within the asp code is known to
follow different paths on different system platforms, which can
potentially explain the discrepancy in the niter accumulated on Linux vs
Mac.

Removing this test allows it to be run on a Mac (rather than skipped),
and given that the final image values are within tolerance on all
systems it indicates that the images are scientifically equivalent
despite the niter being different.