CAS-13901 : Skip asp tests on Mac darwin
Given the difficulty in nailing down the niter criterion on Mac M1, it
was decided to continue skipping the tests on Mac and add a disclaimer
to the documentation about running asp on Mac.

The long term solution is to find a more stable minimizer than LBFGS
that yields the same numbers on different platforms.