CAS-14517: Add more importfitsidi tests
A recent casacore update brought in a fix for a bug that was
triggered when data from different VLBI correlators was imported
within the same CASA session.  Test that this works correctly now.
Also add a test for importing VLBA archival data that was
correlated with the old VLBA hardware correlator, which was blocked
by the aforementioned bug.